.github/scripts/issue-management/python_tool/requirements.txt
pygithub>=1.58.0
click>=8.1.3
pyyaml>=6.0
requests>=2.28.1
python-dotenv>=0.21.0
